+++ changelog.rst
@@ -6,6 +6,7 @@ Change Log
 
 - 2021-12-10
   - Include leading white-space to properly align indented multi-line doc-strings.
+  - Add ``doc-show-inline-exclude-regexp`` to optionally exclude comments matching a regular expression.
 
 - 2021-12-08
   - Keep blank space between the symbol and comments.